[models][m]: auth models renamed to jwt as that is what is it for. tests mocks and imports updated - refs #401
