+++ app/__init__.py
@@ -20,7 +20,7 @@ from werkzeug.exceptions import NotFound, Unauthorized, MethodNotAllowed, BadReq
 from .database import db
 from .schemas import ma
 from app.auth.controllers import auth_blueprint, bitstore_blueprint
-from app.auth.models import JWT
+from app.auth.jwt import JWT
 from app.logic.db_logic import get_user_by_id
 from app.package.controllers import package_blueprint
 from app.site.controllers import site_blueprint
