+++ app/auth/annotations.py
@@ -10,7 +10,7 @@ from flask import current_app as app
 from flask import request, _request_ctx_stack
 
 from app.utils import InvalidUsage
-from app.auth.models import JWT
+from app.auth.jwt import JWT
 from app.auth.authorization import is_authorize
 from app.logic import db_logic
 from app.package.models import Package
